MOOLYA-2250 and MOOLYA-3170 are fixed

+++ server/MlExternalUsers/externaluser/externalUserSchema.js
@@ -186,8 +186,8 @@ let externalUser = `
       addressCountry : String
       addressCountryId : String
       addressPinCode : String
-      latitude        :  Int
-      longitude       :  Int
+      latitude        :  Float
+      longitude       :  Float
       isDefaultAddress:Boolean
     }
