+++ server/MlInternalUsers/admin/registration/registrationSchema.js
@@ -34,8 +34,8 @@ let registrationSchema = `
         addressCountry : String
         addressCountryId : String
         addressPinCode : String
-        latitude        :  Int
-        longitude       :  Int
+        latitude        :  Float
+        longitude       :  Float
         isDefaultAddress:Boolean
      }
      
@@ -103,8 +103,8 @@ let registrationSchema = `
       addressCountry : String
       addressCountryId : String
       addressPinCode : String
-      latitude        :  Int
-      longitude       :  Int
+      latitude        :  Float
+      longitude       :  Float
       isDefaultAddress:Boolean
     }
